Order of plot affects shared axes' limits
Two variables that have a different range
are plotted in shared axes.

In one example, the order of plot has no effect
on the range on the axes.

In another, the axes
range is different for different order of plot.
It appears that the range is determined by the
range of the variable plotted on the LAST axis
in the figure.