Check if a method is defined in derived class
How does the author of a base class ensure that a method
called by the base class is defined in the derived class,
without access to the code of the derived class?

This commit uses the __build_class__ function, which
corresponds to the top-level `class` function, for building
a class in Python.  The function can be modified by the
author of the base class, so that, when the base class'
module is imported into the derived class' module,
`class` checks whether the said method is defined in the
derived class.